Move tests into another submodule.
